fix: Make sure IntellIJ community verifier passes
