+++ src/main/resources/META-INF/plugin.xml
@@ -570,8 +570,6 @@
         <xml.xmlExtension implementation="de.timo_reymann.mjml_support.xml.MjmlXmlExtension"
                           order="first"/>
 
-        <psi.referenceContributor language="mjml"
-                                  implementation="de.timo_reymann.mjml_support.reference.MjmlPathAttributeTypeReferenceContributor"/>
         <completion.contributor language="mjml"
                                 order="first"
                                 implementationClass="de.timo_reymann.mjml_support.completion.MjmlPathAttributeTypeCompletionContributor"/>
@@ -595,11 +593,7 @@
                          enabledByDefault="true"
                          groupName="MJML"
                          displayName="Unknown attributes in mjml tags"/>
-        <localInspection implementationClass="de.timo_reymann.mjml_support.inspection.InvalidPathAttributeInspection"
-                         language="mjml"
-                         enabledByDefault="true"
-                         groupName="MJML"
-                         displayName="Invalid value for mjml attribute of type path"/>
+
         <localInspection implementationClass="de.timo_reymann.mjml_support.inspection.InvalidPixelAttributeInspection"
                          language="mjml"
                          enabledByDefault="false"
